# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ConversationInput {
  /** @maxLength 120 */
  title?: string;
}

export interface Conversation {
  id: string;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface Message {
  id: string;
  role: MessageRole;
  content: string;
  createdAt: string;
  /**
     * @minimum 0
     * @maximum 4
     */
  signalLevel?: number;
}

export type ConversationDetail = Conversation & {
  messages: Message[];
};

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 12000
     */
  content: string;
}

export interface SafetySignals {
  confirmationBias: number;
  emotionalDependency: number;
  anthropomorphism: number;
  delusionReinforcement: number;
  escalatingDistress: number;
  unsafeAdvice: number;
  overValidation: number;
  manipulation: number;
  hallucinationRisk: number;
}

export type EnvironmentStateMode = typeof EnvironmentStateMode[keyof typeof EnvironmentStateMode];


export const EnvironmentStateMode = {
  grounding: 'grounding',
  reflective: 'reflective',
  depth: 'depth',
  growth: 'growth',
  complexity: 'complexity',
} as const;

export interface EnvironmentState {
  mode: EnvironmentStateMode;
  approvedObjects: string[];
  /** @nullable */
  suggestedObject: string | null;
}

export interface MessageResult {
  userMessage: Message;
  assistantMessage: Message;
  signals: SafetySignals;
  highestSignal?: string;
  /**
     * @minimum 0
     * @maximum 4
     */
  interventionLevel: number;
  environmentState: EnvironmentState;
}

export interface TimelinePoint {
  label: string;
  detail: string;
}

export interface Reflections {
  topics: string[];
  perspectives: string[];
  questions: string[];
  timeline: TimelinePoint[];
}

export type EnvironmentUpdateDecision = typeof EnvironmentUpdateDecision[keyof typeof EnvironmentUpdateDecision];


export const EnvironmentUpdateDecision = {
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface EnvironmentUpdate {
  objectName: string;
  decision: EnvironmentUpdateDecision;
}

export interface ConversationAnalytics {
  total: number;
  active: number;
  messages: number;
  averageDuration: string;
}

export interface SignalPoint {
  label: string;
  confirmationBias: number;
  dependency: number;
  anthropomorphism: number;
  distress: number;
}

export interface InterventionPoint {
  label: string;
  level0: number;
  level1: number;
  level2: number;
  level3: number;
  level4: number;
}

export interface ModelAnalytics {
  provider: string;
  status: string;
  averageLatency: string;
  failures: number;
}

export interface ResearchAnalytics {
  conversations: ConversationAnalytics;
  signalSeries: SignalPoint[];
  interventionSeries: InterventionPoint[];
  model: ModelAnalytics;
}