# lib/api-zod/src/generated/types/environmentState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EnvironmentStateMode } from './environmentStateMode';

export interface EnvironmentState {
  mode: EnvironmentStateMode;
  approvedObjects: string[];
  /** @nullable */
  suggestedObject: string | null;
}