# lib/api-zod/src/generated/types/environmentUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EnvironmentUpdateDecision } from './environmentUpdateDecision';

export interface EnvironmentUpdate {
  objectName: string;
  decision: EnvironmentUpdateDecision;
}